+++ ui/cli/user_menus.py
@@ -6,9 +6,10 @@ Implementation of the specific menus used in the CLI of the program.
 
 Classes:
     MainMenu
+    HelpMenu
 """
 
-__version__ = "0.0.1.0"
+__version__ = "0.0.1.1"
 __date__ = "24-09-2018"
 __status__ = "Development"
 
@@ -46,13 +47,17 @@ class MainMenu(bue.SimpleMenuCLI):
     def onExit(self):
         """
         Handler of the event - 'exit' from the program. Enforces the termination
-        of the program.
+        of the program by exiting the main menu loop, since it sets the menu
+        status to the sudoku_py.ui.cli.basic_ui_elements.DEF_OK_STATUS value.
         
         Signature:
-            None -> None
+            None -> str
+        
+        Returns:
+            str: fixed value 'Exit from the main menu and the program'
         """
         sys.stdout.write('Bye!\n')
-        self._strStatus = bue.DEF_OK_STATUS
+        self.Status = bue.DEF_OK_STATUS
         return 'Exit from the main menu and the program'
     
     def onNewGame(self):
@@ -107,3 +112,60 @@ class MainMenu(bue.SimpleMenuCLI):
                 etc.
         """
         return 'onGeneratePuzzle'
+    
+    def onHelp(self):
+        """
+        Handler of the event - 'show help'. ...
+        
+        Signature:
+            None -> str
+        
+        Returns:
+            str: result of the action initiated by this menu item, i.e.
+                the value returned by the help menu.
+        """
+        return self._launchChild()
+    
+    def onShowRecords(self):
+        """
+        Handler of the event - 'show high scores / records'. ...
+        
+        Signature:
+            None -> str
+        
+        Returns:
+            str: result of the action initiated by this menu item, i.e.
+                the value returned by the help menu.
+        """
+        return 'onShowRecords'
+
+class HelpMenu(bue.SimpleMenuCLI):
+    """
+    Help / information / usage menu of the CLI application.
+    
+    Subclasses sudoku_py.ui.cli.basic_ui_elements.SimpleMenuCLI.
+    
+    Methods:
+        run()
+            None -> str
+    """
+    
+    #class fields
+    
+    _strMenuName = 'Help'
+    
+    #helper methods - event handlers
+    
+    def onBack(self):
+        """
+        Handler of the event - 'back to the previous menu'. ...
+        
+        Signature:
+            None -> str
+        
+        Returns:
+            str: result of the action initiated by this menu item, i.e.
+                'Back from the help menu'.
+        """
+        self.Status = bue.DEF_OK_STATUS
+        return 'Back from the help menu'